Display Bundle Errors in debug console (#964)
This PR displays bundle errors in debug console and focuses on the error
source, because some bundle errors happen before the application is
activated and therefore the cdp debugger we rework how the debug session
is created, we first create just a shell of debugSession, that is not
connected to the debugger, but is able to log messages into debug
console and latter when the application is ready we connect the
debugSession to it.

Because this new pattern introduces a possibility of DAP methods being
called before `debugAdapter` is connected to `cdp` we move cdp related
functionality to debug session and return empty responses in case it
does not exist yet.

### How Has This Been Tested: 

-run `react-native-78` test app and force bundle error also check if the
app is working as expected when syntax is correct.


https://github.com/user-attachments/assets/314df332-c2f8-4a3e-a889-b5171c2d6c79