Rework CI to build fresh versions for publication (#1023)
This PR changes the behavior of the publishing CI to allow for automatic
packaging process to take place when publishing instead of using an
artefakt created for github publication.
This change is needed because `vsce` and `osvx` publishing commands are
not capable of changing the metadata of already packaged `vsix`
directories and the universal target specified in original artefakt
would conflict with targets specified by CI user.

### How Has This Been Tested: 

it can only be tested by running it :(