Remove external publishing workflow form the CI (#1024)
This PR changes the approach to the publishing CI and removes the
external `HaaLeo/publish-vscode-extension@v1` replacing it with
underling `vsce` and `osvx` commands. This is necessary because of a bug
that was surfaced by
[this](https://github.com/software-mansion/radon-ide/actions/runs/13769019110)
CI run.

### How Has This Been Tested: 

The final version was not tested as it would mean running the CI