Support inline shell scripts for custom build script (#1031)
This PR adds a support for inline shell script in
`customBuild.{Platform}.buildCommand` configuration. For many users the
mos intuitive way of utilizing build command was to wirite something
like `echo ${PWD}/apkPath` this would not work tho as `$ENV` syntax
requires full shell, so we start running `buildCommand` in shell.

### How Has This Been Tested: 

build `expo-52-prebuild-with-plugins` android project and find generated
apk in `android/app/build/outputs/apk/debug` copy this apk to the root
of the application and set the android build command as `echo
${PWD}/apkPath` remove the whole android folder from the project root
and run application, it works correct.