Use vscode-js-debug as the debugger implementation (#1001)
Uses `js-debug`'s implementation of the JS debugger for targets which
support it (Hermes with RN version >= 0.76).
This allows us to use the features already implemented there (exception
breakpoints, conditional breakpoints) without having to reimplement
them, and possibly allows us to remove our CDP debugger implementation
in the future, once we drop support for the older RN versions.

We patch `vscode-js-debug` in order to:
- create a new session type used by Radon, based on `js-debug`'s `node`
debug session type
- disable the debug toolbar for that session
- disable the use of proposed vscode extension APIs which prevent the
extension from building

We use the `vscode-cdp-proxy` package for implementing the CDP proxy
sitting between the `js-debug` debugger implementation and the
application.
We patch it to:
- fix a bug which incorrectly treats CDP commands as replies when they
include the `id` field
- fix a bug which exports an nonexistent module from the package

### How Has This Been Tested: 
- launch various apps from `test-apps` repo
- verify that in each app, the debugger still works, and:
- the debugger toolbar doesn't appear (we provide our own controls when
a breakpoint is hit)
  - you can set "caught/uncaught exception" breakpoints and they work
  - setting regular breakpoints work correctly
  - setting conditional breakpoints work
  - logs are routed to the debug console
- pausing inside RN code works correctly (i.e. when handling caught
exceptions)
- running the CPU profiler works and maps the source locations correctly