Refactor DebugSession and DebugAdapter to support only spawning JS debug session. (#1048)
In #964 we changed the debugger workflow and started using the debugger
to also wrap the metro session to allow us reporting bundle errors.

Further #985 and #1001 contributed some changes to the debugger
workflows that impacted the way we handle debug commands with assumption
there is a separate debug session for metro and a new one for the js
debugger.

This PR clears up that assumption and allows for the DebugSession class
to work with the parent metro session but also without it. While the
latter mode is not used, this refactor is needed to build the new
"connect" functionality that allows Radon to work with the externally
controlled metro instances. In such scenario we don't get bundle errors
and hence can't use the debug session to report these. In addition, we
need to handle disconnect events.

In this PR we are introducing a concept of master and JS debug sessions.
Previously we'd use a different naming that I find confusing. The master
session is the one that is spawned immediately along metro and is used
to push errors from the bundler. The js debug session is started as a
child session of the master session and is the main one that controls
the JS runtime. The important part is that the JS debug session can now
also launch without the parent session.

Below are highlights of what we are changing as part of this refactor:
- DebugAdapter is no longer the owner of the js debug session. Instead,
DebugSession is the main class that controlls both parent and child
session. This removes some complexity from DebugAdapter and allows to
avoid the adapter from acting as a command proxy
- We update ping implementation introduced in #985 to use response
instead of events for communication. This simplifies the implementation
and event handling.
- We extract a logic for starting debug session that also captures the
session reference. This shared logic is now used to launch the parent
metro session and the JS debugger session.
- We rename the session that we start for the names to better indicate
those sessions are spawned and controlled by Radon IDE
- We add a new onTerminate event that is triggered when js debug session
is disconnected.
- We move call to `DebugSession.start` from DeviceSession constructor
into `start` method. This allow us to await the start of the parent
session before we continue with the normal flow. Before we'd rely on the
parent session being started before we report first errors but we'd
never await it.

Fixes # (issue)

### How Has This Been Tested: 
1. Run RN 78 and Expo 52 to test new debugger setups. Test reload JS
functionality
2. Run RN 75 and Expo 51 to test old debuger setup
3. Run Expo Go and test JS reloads to make sure changes from #985 are
still in place (see test scenario from #985)