Properly trigger resume events to support Overlay CDP methods (#1059)
This PR adds support for built-in Overlay (via CDP) that displays
debugger paused status. This only applies to the new debugger.

While the code for supporting this API is added, there is no place in
the codebase currectly that'd use this.

As a part of this change however, we had to refactor support for
paused/resume events.
Before this change, we'd only fire resume event when debugger runs the
Debugger.resume command.
This wasn't ideal as there are other ways how debugger can resume
including the Overlay that displays resume button.

The obvious solution would be to rely on debugger's `Debugger.resumed`
event.
This event however comes also when we perform step-over debugging. 
This causes the overlay to blink for a fraction of a second as the
debugger is resumed to perform a single step.

To workaround this, we're delaying emitter from triggering resumed event
in the case we perform step-over.
Otherwise, we trigger the event immediately.
In case the debugger takes longer to pause after step over, we will
still trigger resume event, but this is very unlikely and it still seems
like a better choice to dismiss the overlay intermittently then to get
the overlay completely out of sync with the debugger state.

### How Has This Been Tested: 
1. Use RN 78 example
2. Set breakpoint, test step-over functionality. Verify the overlay
doesn't "blink" when stepping over and it always gets dismissed.