fix EAS in launch config not being set correctly in webview (#1063)
Fixes an issue where the frontend was never aware of the launch config's
EAS settings.
This was caused by the frontend and extension using a different method
to read the launch config, one of which picked the fields from the
config to send explicitely, and which omitted the `eas` entry.

### How Has This Been Tested: 
- open the `expo-52-eas` app
- go to launch config options
- verify the development profile is picked for EAS build
- try to build the app after changing something which impacts the
fingerprint (to force build fail)
- verify an EAS-related error is displayed