Decouple metro and debug session classes (#1066)
This PR cleans up the code preparing for changes that'd allow Radon to
use existing metro.

In this PR we are decoupling debug session to no longer take metro as
argument. This is needed such that we have better control over debug
session w/o need of maintaining up-to-date instances of metro class.
Eventually, we may want to refactor metro class entirely such that it
has a cleaner interface as now it mutates it states when it searches for
debugger URL which can be a bit misleading. Regardless – this is out of
the scope for this change as we only extract debug session such that it
can work based on the session configuration as opposed to fetching
details via metro instance.

What's changed:
1. We simplify and move reconnect logic to device session class. Instead
of doing promise race we just list the endpoints first. It doesn't seem
like the complexity of that code is justified given listing metro
endpoints only takes a few milliseconds. The logic is now much simpler
as we fetch the endpoints first, and only when our old endpoint is
listed we perform ping to check if it is alive. If reconnection is
needed we just start a new JS debugger session and debugSession handles
that correctly.
2. We extract startDebugging helper method to separate file such that it
can be used by ProxyDebugAdapter – this also simplifies the logic as the
adapter class no longer needs to have try/finally block and get the
session instance via the callback – this is something that the helper
method we added recently takes care of.

The main goal of this change was to be able to delete metro import from
debugSession file.

### How Has This Been Tested: 
1. Test new debugger connection work as expected including JS reloads
(tested on RN78 example)
2. Test reloads on expo go app – make sure debugger stays connected
after reloads.