Pass build errors from Extension to Webview (#1068)
Refactors how build errors are reported to the user.
Previously, the error message shown in the error dialog would be
generated in the webview code, based on the state of the workspace and
launch configuration.
With this PR, the error is passed by the extension code to the webview
for display, along with some extra context (platform and build "type")
that the webview needs to adjust the actions available to the user.
This way, we can generate the error message close to where the error
actually occurs instead of essentially "guessing" the cause at the time
we want to display it.

### How Has This Been Tested: 
Ran `expo-52-eas` and `react-native-78` test apps, breaking configs:
- `app.json` in `expo-52-eas`
- remove `ios` and `android` directories in `react-native-78`
- uninstall `node` and relaunch IDE
To check that relevant errors appear.