Adapt updated sourceMapOverrides format in old debug adapter (#1071)
This PR updates the original debug adapter implementation to support the
updated format of source map aliasing.

For the new react native debugger, we use the option
`sourceMapOverrides` in order to translate from "/[metro-project]/"
alias that the RN devtools introduced. With the old debug adapter, we've
been expecting a different format for this, namely, a list of pairs with
alias and absolute file location. Given we may want to use the old debug
adapter to support the new debugger while keeping an option to rely on
vscode-js-debug, we want to unify these formats, such that we don't need
to format the parameters differently depending on the debug adapter we
use.

This PR adapts ths old debugger implementation to accept the updated
format. Now the overrides are provided as `Record<string,string>` and
the override patterns may include wildcards. The basic implementation
only works with wildcards at the very end of the path, but given we
currently generate the patterns on our own and they meet these criteria
this seems like a good tradeoff given the complexity of different
pattern schemas we'd have to support for full compatibility.

### How Has This Been Tested: 
1. Open app with new debugger (RN 78 for example)
2. In DebugSession change the line: `const debuggerType =
OLD_JS_DEBUGGER_TYPE;` to force the old proxy adapter to be used
3. Test breakpoints, logs etc