Handle call frames when logs don't use radon's runtime (#1072)
When connecting to the new RN debugger without Radon's runtime, the logs
wouldn't appear in the debug console.

The reason was that we improperly transform the output event because we
had a special path for creating events that only works when extra
parameters are provided via our `wrapConsole` override from
`runtime.js`.

However, when we connect to existing metro instance, the override
doesn't work, yet we still need to handle the console events. There were
a few problems with that:
1) In #846 we changed the way we handle variable references. This was
updated in the main path that relied on the call frame being provided
from the runtime but not in the `else` branch.
2) When source frame isn't provided via runtime, we never attempt to
read the call frame from CDP call which now is properly populated in the
new RN debugger.
3) Finally, when connecting to the new debugger, upon connection is
sends a number of console API calls from before the debugger was
connected. Apparently, before we have the source maps processed we are
unable to properly decipher the stack frame.

We do the following to address these problems:
1) We update the `else` branch in CDPDebugSession where no stack
information is provided in the console api call. We basically just need
to delete the manual handling of the variable reference as it was done
for the main branch in #846
2) We adapt `skipFiles` option and when processing
`message.params.callFrame` from CDP message we filter the first frame
that doesn't match the skipFiles list.
3) To addess the last issue, we delay handling of console API calls
until we get "debug ready" event.

### How Has This Been Tested: 
1. This can only be change with the new Radon connection functionality
that will publish a diff off as a follow up
2. The test relies on connecting to existing metro instance and seeing
the logs with proper stack frame links