Add options to trigger home button and appswitcher buttons (#1081)
This PR adds options in device settings menu that can trigger home and
appSwitch button presses on active device.

We utilize the recently updated sim-server for this purpose that now
supports new `button` command.

While it may not be the optimal place to add such buttons, they aren't
used as frequenly, so we can alter this location once we get more
feedback from the users. On top of that we're setting a shortcut that is
compatible with iOS simulator for the home button press: Cmd+Shift+H

This change refactors the Project interface and removes method for
biometric auth in favor of more generic method that allows us to run a
command. This way we can hook into the commands API easily w/o expanding
this interface further.

Here's how the new options look like in the device settings menu:
<img width="318" alt="image"
src="https://github.com/user-attachments/assets/3d410565-cb30-45b5-91cd-f46edd071c4e"
/>

### How Has This Been Tested: 
1. Run Android, test home and app switch
2. Do the same on iOS