First iteration for the Radon Connect functionality (#1079)
Radon Connect aims to ring support for classic workflows when people
start metro/expo on their own and launch the app on device or existing
simulator.

This PR adds the ground work for this functionality that will currently
only be available in the extension debug builds while we figure out some
UX/DX around it to be able to enable it for everyone.

This PR should have no effect on the builds that we publish, but will
allow us to work on the "connect" functionality without worrying of the
branch getting outdated.

Here's what we are adding:
1) We're adding a new StatusBarItem to the editor status bar that will
be used to control connect functionality. When the StatusBarItem is
shown we hide the "Open Radon IDE" button from the editor navbar. This
is done via the new context variable `RNIDE.showsStatusBarItem` that is
only set when connect functionality is enabled (in development builds
only)
2) When connect is enabled, it scans for metro instances on ports
8081,8082,8083 (port can also be customized). When instance is detected
and it belongs to the current workspace, we connect to the first device
that appears on the list.
3) For now, we only connect the JS debugger. The debugger supports the
new overlay API such that when paused, you can use the device overlay to
resume.
4) We extract interface for Metro and move most of the method to a
MetroLauncher class that implements this interface. We refactor the
extension to use MetroLauncher which is now responsible for starting and
managing metro process, while Metro interface can be used as an
interface for interact with either launched or existing metro process.

In a follow up we plan to work on the following aspects that are
currently not resolved:
 - [ ] connecting plugins like network/redux/etc
 - [ ] metro port customization
- [ ] add some UI to the main panel to indicate when "connect" is active
and to avoid using both the panel and "connect" functionality
- [ ] figure out how to best handle lack of debugger controls (we
currently hide debug controls)

### How Has This Been Tested: 
1) start rn78 project from terminal (`npm run ios`) and have the
extension open.
2) Notice the status bar item changes from "no process running" to "no
device connected" to "connected"
3) Once connected test breakpoints and logs (jump from console to
source)