legal: add Data Processing Addendum & update Privacy Policy (#1086)
This PR adds a new `/legal/dpa` page. It allows users to sign a DPA with
Software Mansion considering the Radon IDE and the newest Radon AI
feature.

Also, updated the Privacy Policy and added a link to DPA on the `/legal`
page.

The page leads to a a .pdf file hosted on the Radon IDE Portal bucket
storage. The page looks as follows:


![image](https://github.com/user-attachments/assets/2df35b54-529d-4a95-96fb-c3e28bcb8a71)

## How to test?

Visit the Vercel Preview of our website or run the docs locally