Add launch arguments as a configuration on iOS. (#1096)
This PR adds a launch arguments to launch configuration on iOS, for now
it only works with non expo applications, but as we'll move away from
utilizing expo deep links we can add it to them as well.

This configuration is useful for projects that use firebase and for full
debugging need to add `-FIRDebugEnabled` argument to launch command.
https://firebase.google.com/docs/analytics/debugview

### How Has This Been Tested: 

1) add this configuration to `react-native-79` test app and check if it
breaks anything
2) breakpoint in `launchApp` method and see if the argument is passed as
expected
3) watch the preview of the documentation on vercel