Refactor build type selection logic to happen outside the builders (#1097)
Previously, the "builders" (`buildIos` and `buildAndroid` functions)
were responsible for selecting the "type" of the build they should
perform, based on the radon Launch Configuration and the application
workspace. This resulted in a lot of duplicated logic between these
functions and makes it impossible for the callers of these functions to
control which type of build should be run.

This PR changes these functions to accept a `BuildConfig` parameter
which specifies the build type instead, and moves the build type
selection logic to the `BuildManager`.

## Test plan:
- open test apps from the test app repo and verify the correct build
type is used