Prepare DeviceSession management for multidevice capability  (#1084)
This PR is an initial ground work that will allow as to keep devices
alive even when another one is selected.
Keep in mind that this functionality is never actually used as the
current debugSession and tools structure wouldn't work with new setup,
but adding those changes a part of this PR would make it even bigger and
would slow down review process significantly.

### What has been done:
1) the `metro` and `devtools` objects are now part of `deviceSession`
class, so we can create as many of them as we need
2) `DeviceSessionsManager` class was created and it is responsible for
selecting and managing running devices.
3) restart/reload system was collapsed to a single `reload` method with
additional options, it was necessary as we no longer can rely on
`selectDevice` as a mechanism for reseting `deviceSession`
4) deviceSession start and restart systems are now updated to use new
`cancelToken` implementation that allows as to cancel in any async
methods and not only childProcesses.
5) We add `reboot` method to the deviceBase interface, before this
refactor we achieved the same effect with disposing the device and
creating a new session entirely, but as we don't wont to do that anymore
we need to introduce a reboot functionality.

### What more needs to be done to add true multi device functionality?
- multiple debugSession management
- multiple tools management 
- UI for keeping devices alive

### How Has This Been Tested: 

- run `react-native-78` test app and switch between devices 
- use every reload option and do it even if the old reload haven't
finished
- run `documents` folder as a root of the workspace and switch between
some application by selecting `appRoot` in launch configuration
- run `react-native-77` test app and remove android directory to check
if native changes detected functionality works as it too was moved from
the project, which made it easier to decouple device session from
project.