Use official react devtools package for handling devtools communication (#1110)
This PR updates devtools.ts class to rely on devtools bridge and store
from the official react's package.

The reason is that we want to start using things like react profiler
that requires a proper communication channel being set up for it to
understand all the details about the renderer.

We use a custom "headless" build of react-devtools that is placed here:
https://github.com/software-mansion-labs/react/tree/radon-ide/packages/react-devtools-radon-ide
– the reason why we have a custom build is that what we'd actually like
to use is `react-devtools-inline` package, however, the "devtools
frontend" of that packages has a bundled in devtools components and CSS
code. Since we run the "frontend" code in the extension's node
environment, the inline package fails to load as it expects `document`
and `window` to be present due to the fact it bundles in the
component/dom code.

Since we don't expect frequent updates of the devtools frontend package,
and the fact it is relatively large repo that includes all the react
code, we don't pull it in as submodule, but instead just rely on the
build being performed in the radon-ide speciifc package in the repo and
the artifacts to be cloned into third-party directory under the
extension package here.

Apparently, the interface of the devtools bridge only enables interface
for handling events by a specified name. There is no way to register for
"all" types of event dispatched via the devtools channel. Apparently, we
relied on this mechanism in several places across the codebase.

To adapt to this new approach, we explicitly list all the event types
and refactor the listeners to only register for events for that specific
name. We also update the registration hooks to return disposables as it
integrates with the `dispose` mechanism we already use in many places.

On top of that, the devtools class, has to register for all the custom
event types when the bridge is created. This allows it to capture all
the events and dispatch them to the right listener. Note that in our
codebase we sometimes may register for an event, before the bridge is
setup.

Since we adapt a new communication scheme that has a fixed list of
events, we are refactoring plugins (network/redux/react query) to use a
single message type with a unique scope parameter. This allows us to
avoid coupling devtools with every plugin that we have or want to add in
the future. As part of this refactoring we're migrating:
1. Network plugin to use RNIDE_pluginMessage for back and forth
communication. We also delete separate message types for enable/disable
events and only rely on CDP's Network.enable/disable messages
2. We update redux devtools to also use the standard plugin message type
"RNIDE_pluginMessage" with "redux-devtools" scope. We rename the plugin
ID to match the naming convention we used for other plugins and strip
"RNIDE" prefix.
3. Finally, we update the react query plugin to also use the same
"RNIDE_pluginMessage" type for communicating over devtools channel.

### How Has This Been Tested: 
1. Run test app with plugins
4. Verify network/redux/etc plugins work as expected
5. Check switching between devices

---------

Co-authored-by: Juliusz Wajgelt <49338439+jwajgelt@users.noreply.github.com>