Fix device sesstion start (#1115)
A last minute addition I made in #1110 introduced a regression where the
device session fails to start. The reason was that I extracted code that
was meant to control creating devtools instance, but it should not rely
on the devtools instance already set on `this` but rather the new
devtools instance created in that method.

### How Has This Been Tested: 
1. Start any project, see the panel launches normally. Before the panel
would get start on 'initializing device" step.