Add support for expo go in expo 53 (#1123)
In RN 77 there were changes to how runtimes descriptions are generated
that we addressed in #864. As the new expo go on expo-53 is using a
newer version of react native those changes where interfering with how
we recognize that the runtime is expo-go, this PR fixes that by
additionally using `appId` filed missing from older versions.

### How Has This Been Tested: 

-Run `expo-53` test app in expo go mode (without prebuilt) and test
debugger on ios