fix debugger always restarting after app reload (#1120)
Currently, the logic for reconnecting the debugger after the application
is reloaded always restarts the debug session.
This was uncaught since the behaviour is _technically correct_ (in that
in the end, we'll always have a debug session connected to the
application after reload), but undesirable (we'd rather reuse the
existing debug session since its initialization is relatively slow).
This PR fixes the minor bugs which caused the issue.

Right now, after this fix is applied, the current debugger
implementation breaks after JS reload, presumably due to issues with how
source maps are cleaned up.
This surfaces by not mapping the sources correctly in the Debug Console
(instead always mapping to `__source__`)
This PR also makes the debug session always restart on the old debugger
as a workaround for that issue.

### How Has This Been Tested: 
- use the `PROXY_JS_DEBUGGER_TYPE` in `DebugSession.startJSDebugSession`
to enable the new `js-debug`-based debugger
- add log points to `DeviceSession.reconnectJSDebuggerIfNeeded`
- check that after "Reload JS" is triggered in Radon, the function
returns early (without calling `this.connectJSDebugger`)